increased soft limit of pixels on 8266 to 150

+++ controller/effects.h
@@ -1,4 +1,4 @@
-void walk(uint32_t(*readPixel)(uint8_t), void(*setPixel)(uint8_t, uint32_t, boolean), uint8_t start, uint8_t end) {
+void walk(uint32_t(*readPixel)(uint16_t), void(*setPixel)(uint16_t, uint32_t, boolean), uint16_t start, uint16_t end) {
       int firstPixel = (*readPixel)(start-1);
       for (int i=start; i < end; i++) {
         (*setPixel)(i-1, (*readPixel)(i), false);
